fix(proximity): resolve cross-world misidentification and preserve accurate native instance locations
- Update FriendRepository.UpdateFriendFromCharacter to inherit CurrentWorldId from the rendering LocalPlayer, bypassing unreliable remote character world data
- Implement IsStandardTerritory in GameDataService to distinguish open-world environments from private instances (e.g., housing, islands)
- Inject IGameDataService into ProximityService to conditionally prevent overriding highly accurate native location IDs with generic unresolvable instance IDs
- Update respective unit tests to satisfy newly injected dependencies and validate local player world extraction logic

Resolves proximity-induced location masking and world name fallback bugs

+++ Befriender/Core/Friends/Contracts/IFriendRepository.cs
@@ -11,7 +11,7 @@ public interface IFriendRepository {
 
     IReadOnlyList<FriendProfile> GetFriends();
     void UpdateFriends(IEnumerable<FriendProfile> friends);
-    void UpdateFriendFromCharacter(ulong contentId, IPlayerCharacter player, ushort territoryId);
+    void UpdateFriendFromCharacter(ulong contentId, IPlayerCharacter player, uint territoryId);
     void RemoveFriendData(ulong contentId);
     void Save();
     void ClearCache();
